% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalroi.R
\name{BlockPermu}
\alias{BlockPermu}
\title{Block permutation of matrix}
\usage{
BlockPermu(X, BSize)
}
\arguments{
\item{X}{A matrix representing the channel of image/matrix.}

\item{BSize}{An integer that is the size of permuted block.}
}
\value{
The matrix after block-wise permutation.
}
\description{
\code{BlockPermu} permutes single channel of image/matrix block-wisely.
}
\details{
This function can randomly permute the matrix block-wisely.
}
\examples{
X=matrix(runif(64*64),64,64)
BlockPermu(X,8)

}
\seealso{
\code{\link{colocalroitest}}
}
\author{
Shulei Wang
}
