% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalpixel.R
\name{colocalpixel}
\alias{colocalpixel}
\title{Spatial Adaptive Colocalization Analysis}
\usage{
colocalpixel(X, Y, alpha = 0.05, Thresholds = NULL, method = "bonferroni")
}
\arguments{
\item{X}{A numerical matrix, the intensity matrix of the first channel/red channel.}

\item{Y}{A numerical matrix, the intensity matrix of the first channel/green channel.}

\item{alpha}{An numerical value, the type I error in multiple testing.}

\item{Thresholds, }{A vector of length 2, the input threshold for each channel when Manders' split colocalization coefficients and maximum truncated Kendall tau correlation coefficient are evaluated.}

\item{method}{A string, indicating which multiple testing is used. Now only \code{'bonferroni'} is supported.}
}
\value{
\code{colocalpixel} returns an object of class \code{colocalpixel}. 
 \code{\link{plot}} can be used to plot image with identification result.
 \code{\link{HeatMapZ}} can be used to plot heat map of Z values.
 
An object of class \code{colocalpixel} is a list containing following components:
\itemize{
\item \code{Zvalue} a numerical matrix to store the Z values of each pixel
\item \code{Pvalue} a numerical matrix to store the P values of each pixel
\item \code{SigPixel} a binary matrix to store the indicator of significance of each pixel
\item \code{Red} the original Red channel
\item \code{Green} the original green channel
}
}
\description{
\code{colocalpixel} is used to analyze colocalization locally and identify correlated region
in dual-channel image.
}
\details{
This function is based on new spatial adaptive colocalization analysis framework, which is able to quantify the degree of colocalization at each pixel level.
}
\examples{
X=matrix(runif(32*32),32,32)
Y=matrix(runif(32*32),32,32)
colocalpixel(X,Y)

}
\seealso{
\code{\link{HeatMapZ}}

Other spatial colocalization analysis: \code{\link{HeatMapZ}},
  \code{\link{plot.colocalpixel}}
}
\author{
Shulei Wang
}
