% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalplot.R
\name{colocalplot}
\alias{colocalplot}
\title{Colocalization Analysis Plot}
\usage{
colocalplot(X, Y, mask = NULL, method = "scatter")
}
\arguments{
\item{X}{A numerical matrix, the intensity matrix of the first channel/red channel.}

\item{Y}{A numerical matrix, the intensity matrix of the first channel/green channel.}

\item{mask}{An logical matrix, where TRUE means pixel is involved into analysis and vice versa. Dimension of mask must be matched with one of image.}

\item{method}{A string, indicating which plot is made. The plot can only be from \code{c('scatter','li')}.}
}
\value{
Invisible \code{NULL}
}
\description{
\code{colocalplot} is used to make colocalization analysis plot, including scatter plot and Li's plot. 
Scatter plot is a plot where x axis is pixel intensity of green channel and y axis is pixel intensity of red channel
Li's plot provide two plots, the common x axis is the product value and y axis of two plots are pixel intensity of red and green channel, respectively.
}
\details{
This function provides the colocalization analysis plot via plot function in \code{ggplot2}.
}
\examples{
X<-matrix(runif(10*10),nrow=10,ncol=10)
Y<-matrix(runif(10*10),nrow=10,ncol=10)
colocalplot(X,Y)

}
\references{
Li, Qi, et al. "A syntaxin 1, Gαo, and N-type calcium channel complex at a presynaptic nerve terminal: analysis by quantitative immunocolocalization." Journal of Neuroscience 24.16 (2004): 4070-4081.
}
\seealso{
\code{\link{DualImageVisual}}
}
\author{
Shulei Wang
}
