% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalroi.R
\name{colocalroi}
\alias{colocalroi}
\title{Region based Colocalization Analysis}
\usage{
colocalroi(X, Y, mask = NULL, Thresholds = NULL, method = "pearson")
}
\arguments{
\item{X}{A numerical matrix, the intensity matrix of the first channel/red channel.}

\item{Y}{A numerical matrix, the intensity matrix of the first channel/green channel.}

\item{mask}{An logical matrix, where TRUE means pixel is involved into analysis and vice versa. Dimension of mask must be matched with one of image.}

\item{Thresholds, }{A vector of length 2, the input threshold for each channel when Manders' split colocalization coefficients and maximum truncated Kendall tau correlation coefficient are evaluated.}

\item{method}{A string, indicating which plot is made. The plot can only be from \code{c('pearson', 'kendall', 'spearman', 'icq', 'mandersM1', 'mandersM2', 'trunkendall')}.}
}
\value{
The value of colocalization measure.
}
\description{
\code{colocalroi} can evaluate a statistical quantity for colocalization/correlation in a given region. The colocalization measures \code{colocalroi} support include 
Pearson correlation coefficient, Manders' split colocalization coefficients, Spearman’s rank correlation coefficient, Kendall tau correlation coefficient, intensity correlation quotient and maximum truncated Kendall tau correlation coefficient, a robust colocalization measure.
}
\details{
The Manders' split colocalization coefficients and maximum truncated Kendall tau correlation coefficient need the input of thresholds. If it is missing, the default thresholds are caculated by Otsu's method.
}
\examples{
X=matrix(runif(32*32),32,32)
Y=matrix(runif(32*32),32,32)
colocalroi(X,Y)

}
\references{
Manders, E. M., et al. "Dynamics of three-dimensional replication patterns during the S-phase, analysed by double labelling of DNA and confocal microscopy." Journal of cell science 103.3 (1992): 857-862.

Manders, E. M. M., F. J. Verbeek, and J. A. Aten. "Measurement of co‐localization of objects in dual‐colour confocal images." Journal of microscopy 169.3 (1993): 375-382.

Adler, J., S. N. Pagakis, and I. Parmryd. "Replicate‐based noise corrected correlation for accurate measurements of colocalization." Journal of microscopy 230.1 (2008): 121-133.

Li, Qi, et al. "A syntaxin 1, Gαo, and N-type calcium channel complex at a presynaptic nerve terminal: analysis by quantitative immunocolocalization." Journal of Neuroscience 24.16 (2004): 4070-4081.

French, Andrew P., et al. "Colocalization of fluorescent markers in confocal microscope images of plant cells." Nature protocols 3.4 (2008): 619.
}
\seealso{
\code{\link{colocalroitest}}
}
\author{
Shulei Wang
}
