% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalroi.R
\name{colocalroitest}
\alias{colocalroitest}
\title{Statistical Test for Region based Colocalization Analysis}
\usage{
colocalroitest(X, Y, mask = NULL, Thresholds = NULL, method = "pearson",
  bsize = 1, times = 100, is.parallel = FALSE, numcore = 2)
}
\arguments{
\item{X}{A numerical matrix, the intensity matrix of the first channel/red channel.}

\item{Y}{A numerical matrix, the intensity matrix of the first channel/green channel.}

\item{mask}{A logical matrix, where TRUE means pixel is involved into analysis and vice versa. Dimension of mask must be matched with one of image.}

\item{Thresholds, }{A vector of length 2, the input threshold for each channel when Manders' split colocalization coefficients and maximum truncated Kendall tau correlation coefficient are evaluated.}

\item{method}{A string, indicating which plot is made. The plot can only be from \code{c('pearson', 'kendall', 'spearman', 'icq', 'mandersM1', 'mandersM2', 'trunkendall')}.}

\item{bsize}{An interger, the size of permuted block.}

\item{times}{An interger, the times of permutation for null distribution approximation}

\item{is.parallel}{A boolean value representing if the permutation is done by parallel computation}

\item{numcore}{An integer which specifies the number of core in parallel computation if is.parallel is TRUE.}
}
\value{
An object which belongs to the class colocalroitest. 
\enumerate{
\item \code{Index_Value} The observed value of chosen colocalization measure without any permutation.
\item \code{P_value} The P-value calculated from permutation test.
\item \code{Null_Distribution} The value of chosen colocalization measure for each permutation.
}
}
\description{
\code{colocalroitest} can conduct a statistical test for colocalization/correlation in a given region. 
\code{colocalroitest} supports both pixel by pixel and block by block permutation test.
The colocalization measures \code{colocalroitest} support include 
Pearson correlation coefficient, Manders' split colocalization coefficients, Spearman’s rank correlation coefficient, Kendall tau correlation coefficient, intensity correlation quotient and maximum truncated Kendall tau correlation coefficient, a robust colocalization measure.
}
\details{
When \code{bsize} is 1 or mask is given, the pixel by pixel permutation test is conducted. 
When \code{bsize} is larger than 1 and mask is \code{NULL}, the block by blcok permutation test is conducted.
}
\examples{
X=matrix(runif(64*64),64,64)
Y=matrix(runif(64*64),64,64)
colocalroitest(X,Y)

}
\references{
Costes, Sylvain V., et al. "Automatic and quantitative measurement of protein-protein colocalization in live cells." Biophysical journal 86.6 (2004): 3993-4003.
}
\seealso{
\code{\link{colocalroi}}

\code{\link{plot.colocalroitest}}
}
\author{
Shulei Wang
}
