% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalvisual.R
\name{DualImageVisual}
\alias{DualImageVisual}
\title{Dual-Channel Image Visualization}
\usage{
DualImageVisual(X, Y, isSplit = FALSE, isScale = TRUE, mask = NULL)
}
\arguments{
\item{X}{A numerical matrix, the intensity matrix of the first channel/red channel.}

\item{Y}{A numerical matrix, the intensity matrix of the first channel/green channel.}

\item{isSplit}{An logical value. If it's \code{TRUE}, then two channels dispaly individually, and if \code{FALSE}, a image with merging two channels dispaly.}

\item{isScale}{An logical value to determine whether displayed image is scaled.}

\item{mask}{An logical matrix, where TRUE means pixel is involved into analysis and vice versa. Dimension of mask must be matched with \code{X} and \code{Y}.}
}
\value{
Invisible \code{NULL}
}
\description{
\code{DualImageVisual} is used to display dual channel image.
Each channel can be displayed individually or two channels can be displayed after merging together.
This function can also increase contrast ratio of plot by scaling.
If only region of interest is displayed, mask can be used.
}
\details{
This function depends on \code{\link[EBImage]{display}} mainly.
First, image is preprocessed by scaling or masking.
Then, two plots are made when each channel is diplayed individually, and only one is displayed otherwise.
In addition, after scaling, the largest intensity in each channel is always 1.
}
\examples{
X<-matrix(runif(10*10),nrow=10,ncol=10)
Y<-matrix(runif(10*10),nrow=10,ncol=10)
DualImageVisual(X,Y)

}
\seealso{
\code{\link{DualImageSave}}
}
\author{
Shulei Wang
}
