% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalpixel.R
\name{HeatMapZ}
\alias{HeatMapZ}
\title{Heat map of Z value for an \code{colocalpixel} Object}
\usage{
HeatMapZ(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{colocalpixel}}}

\item{...}{Arguments to be passed to methods.}
}
\value{
Invisible \code{NULL}
}
\description{
\code{HeatMapZ} is used to display the heat map of Z values in \code{colocalpixel}.
}
\details{
The positive value of Z value is labelled in blue color and negative value is labelled in red color.
}
\examples{
 
X=matrix(runif(32*32),32,32)
Y=matrix(runif(32*32),32,32)
co <- colocalpixel(X,Y)
HeatMapZ(co)  

}
\seealso{
\code{\link{colocalpixel}}

Other spatial colocalization analysis: \code{\link{colocalpixel}},
  \code{\link{plot.colocalpixel}}
}
\author{
Shulei Wang
}
