% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalpixel.R
\name{plot.colocalpixel}
\alias{plot.colocalpixel}
\title{Plot Identification Result for an \code{colocalpixel} Object}
\usage{
\method{plot}{colocalpixel}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{colocalpixel}}}

\item{...}{Arguments to be passed to methods.}
}
\value{
Invisible \code{NULL}
}
\description{
This function is used to display the dual-channel image in \code{colocalpixel} with identification result.
}
\details{
The significance result are labelled in blue color
}
\examples{
 
X=matrix(runif(32*32),32,32)
Y=matrix(runif(32*32),32,32)
co <- colocalpixel(X,Y)
plot(co)  

}
\seealso{
\code{\link{colocalpixel}}

Other spatial colocalization analysis: \code{\link{HeatMapZ}},
  \code{\link{colocalpixel}}
}
\author{
Shulei Wang
}
