% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalroi.R
\name{plot.colocalroitest}
\alias{plot.colocalroitest}
\title{Plot Permutation Test Result for an \code{colocalroitest} Object}
\usage{
\method{plot}{colocalroitest}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{colocalroitest}}}

\item{...}{Arguments to be passed to methods.}
}
\value{
Invisible \code{NULL}
}
\description{
\code{plot.colocalroitest} makes a plot for permutation test result, including a histgram of null distribution, a red dashed line representing oberserved value and P-value at the top.
}
\details{
The hitogram of null distribution relies on \code{ggplot2}.
}
\examples{
 
X=matrix(runif(64*64),64,64)
Y=matrix(runif(64*64),64,64)
co <- colocalroitest(X,Y)
plot(co)  

}
\seealso{
\code{\link{colocalroitest}}
}
\author{
Shulei Wang
}
