% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI-data.R
\docType{data}
\name{ROI}
\alias{ROI}
\title{A Sample for Region of Interest}
\format{An logical matrix.}
\usage{
data(ROI)
}
\description{
Data is a large boolan matrix indicating the region of interest.
}
\examples{
data(ROI)
}
\keyword{datasets}
